# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ACM-FINTECH",
    "description": "ACM-FINTECH Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/user/update-user": {
      "put": {
        "tags": ["Users"],
        "summary": "Updating user",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the user to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid user supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer-token": []
          }
        ]
      }
    },
...
    "/api/course/update-course": {
      "put": {
        "tags": ["Courses"],
        "summary": "Updating course",
        "description": "This can only be done by the logged in user.",
        "operationId": "updatecourse",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "name of the course to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated course object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid course supplied"
          },
          "404": {
            "description": "course not found"
          }
        },
        "security": [
          {
            "Bearer-token": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer-token": {
      "type": "http",
      "name": "jwt-token",
      "schema": "bearer",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "required": ["email", "password"],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
    "Course": {
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string"
        },
        "isCompleted": {
          "type": "boolean"
        }
      }
    },
...
  }
}